Project_Agent_Zodiac/backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-dotenv
pydantic
requests
